serveur/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
langgraph
unstructured
pillow
unstructured[md]
fastapi[all]